Changed "parseInt()" to "parseFloat()"

+++ code/js/FilterResults.js
@@ -69,8 +69,8 @@ FilterResults.prototype.filterWKT=function(spexresultset){
       s=s.split(" "); // type of s changes from string to array
       //console.log("lat and long split: " + JSON.stringify(s));
       if(s.length!=2) console.log("Error in strArray["+i+"]");
-      longArray.push(parseInt(s[0]));
-      latArray.push(parseInt(s[1]));
+      longArray.push(parseFloat(s[0]));
+      latArray.push(parseFloat(s[1]));
     }
     if(strArray.length!=latArray.length) console.log("Error in latArray!");
     if(strArray.length!=longArray.length) console.log("Error in longArray!");
